Added rake tasks for DB management
This will autoload with Railties if the wcc/rails file is required. It
includes tasks for PostgreSQL and MySQL to dump and restore. A
capistrano task is included to do the restore from the active stage to
the local environment.

Note: I haven't tested the integration of the MySQL tasks I just pulled
them from a project that they worked on before...