Cleaned up the rake tasks and Rails integration
The enhancement to the db:drop task to automatically drop pg connections
before dropping the database is now implemented in a more Rake like way.
Since making this gem appear as an Engine for asset sharing purposes the
Rake tasks were being auto included so I had to restructure some of that
configuration.